+++ internal/source/prepping/other.c
@@ -11,7 +11,7 @@ int Get_Depth(float Number) {
 	return LDE_INVALID;
 }
 
-void Abbreviate_Number(long float Number, char* Buffer, int Size) {
+void Abbreviate_Number(float Number, char* Buffer, int Size) {
 	int Marker = 0;
 	for (int Counter = 0; Counter < LDE_SUFFIXES - 1; Counter++) {
 		if (Number >= 1000) {
@@ -21,7 +21,7 @@ void Abbreviate_Number(long float Number, char* Buffer, int Size) {
 			break;
 		}
 	}
-	int Length = snprintf(Buffer, Size, "%.2Lf", Number);
+	int Length = snprintf(Buffer, Size, "%.2f", Number);
 	if (Core.Suffixes[Marker] != 'J') {
 		Buffer[Length] = Core.Suffixes[Marker];
 		Buffer[Length + 1] = '\0';
